requirements.txt
AWSIoTPythonSDK~=1.4.9
gnureadline~=8.1.2
matplotlib~=3.3
paho-mqtt~=1.5.0
pycryptodome~=3.15.0
pypandoc~=1.5
pysftp~=0.2.9
pytz~=2020.1
requests~=2.28.2
scipy~=1.5
setuptools~=50.3.0
tzlocal~=2.1

scs-core~=3.15.0
scs-host-posix~=3.3.1
